requirements.txt
langchain
langchain-ollama
langgraph
pymupdf
nbformat
nbconvert
jupyter
pydantic
python-dotenv
ipykernel
rich
tavily-python 
duckduckgo-search
ddgs